LayoutTests/fast/css/font-face-insert-link.html

<head>
<script src="../js/resources/js-test-pre.js"></script>
<style>
.test {
  font-family: 'myahem';
}
#a, #b, #container { position:absolute; }
#b { top: 20px }
</style>
</head>

<div id=container>
<div id=a>Text</div>
<div id=b class="test">Text</div>
</div>

<script>
jsTestIsAsync = true;

window.setTimeout(
  function() {
    var link = document.createElement("link");
    link.setAttribute("href", "resources/ahem.css");
    link.setAttribute("rel", "stylesheet");
    link.setAttribute("type", "text/css");
    document.head.appendChild(link);
    window.setTimeout(
      function() {
        var a = document.getElementById('a');
        var b = document.getElementById('b');
        shouldBeFalse('window.getComputedStyle(a).width == window.getComputedStyle(b).width');
        finishJSTest();
      },
     500
    );
  },
  1);
</script>
<script src="../js/resources/js-test-post.js"></script>
