third_party/blink/web_tests/fast/css/font-face-insert-link-expected.txt

PASS window.getComputedStyle(a).width == window.getComputedStyle(b).width is false
PASS successfullyParsed is true

TEST COMPLETE

Text
Text
